requirements.txt
# ===================================================================
# Dependencias del Proyecto: Food Multilabel Classification
# ===================================================================
# Proyecto académico de Machine Learning - Clasificación multilabel de alimentos
# Python 3.8+
# ===================================================================

# Deep Learning & Machine Learning
# -----------------------------------------------------------------
tensorflow>=2.10.0,<2.16.0      # Framework principal de deep learning
keras>=2.10.0                    # API de alto nivel para redes neuronales
scikit-learn>=1.2.0              # Métricas y preprocesamiento ML

# Procesamiento de Datos
# -----------------------------------------------------------------
numpy>=1.23.0,<2.0.0            # Operaciones numéricas y arrays
pandas>=1.5.0                    # Manipulación y análisis de datos
pillow>=9.0.0                    # Procesamiento de imágenes (PIL)

# Visualización
# -----------------------------------------------------------------
matplotlib>=3.6.0                # Gráficos y visualizaciones
seaborn>=0.12.0                  # Visualizaciones estadísticas

# Aplicación Web
# -----------------------------------------------------------------
streamlit>=1.25.0                # Framework para apps web interactivas

# Notebooks
# -----------------------------------------------------------------
jupyter>=1.0.0                   # Notebooks interactivos
notebook>=6.5.0                  # Interfaz de Jupyter Notebook
ipykernel>=6.20.0                # Kernel de Python para Jupyter
ipywidgets>=8.0.0                # Widgets interactivos

# Utilidades
# -----------------------------------------------------------------
tqdm>=4.65.0                     # Barras de progreso
python-dotenv>=1.0.0             # Variables de entorno

# Dataset (opcional - requiere configuración de Kaggle API)
# -----------------------------------------------------------------
kaggle>=1.5.0                    # API de Kaggle para descargar datasets

# ===================================================================
# NOTAS DE INSTALACIÓN:
# ===================================================================
# 
# Instalación básica:
#   pip install -r requirements.txt
#
# Para GPU (NVIDIA CUDA):
#   Instalar CUDA Toolkit y cuDNN primero
#   pip install tensorflow-gpu>=2.10.0
#
# Para desarrollo:
#   pip install -r requirements.txt
#   pip install black flake8 pytest  # Herramientas de desarrollo
#
# Verificar instalación:
#   python -c "import tensorflow as tf; print(tf.__version__)"
#   python -c "import streamlit; print(streamlit.__version__)"
#
# ===================================================================
